Move bucketProportions into nix-support/
On reflection, we don't really want bucketProportions to be in benchmarks/, for
two reasons:

 - We want to get our actual data from a significant number of samples, which
   will take a while (in fact, we want to run for as long as is feasible)
 - We want our benchmarks to run quickly, so we can get feedback and iterate.

These are incompatible. Hence a better place for our eventual bucketProportion
results is release.nix, since that's a fine place for long-running builds to
live. We shouldn't have too much logic in release.nix though: we should keep the
bucketProportions definitions in nix-support/ and just call them release.nix,
with sufficiently large numbers (e.g. 100).

This way, we can still include these proportions in our benchmarks: but we can
measure them using only a few samples.

In particular, we can keep moving scripts out of bucketProportions and into
benchmarkingCommands as needed to keep an eye on performance.