Port getGroundTruths inner loop to Haskell
The tests are passing so it seems to work. This doesn't improve our big-O time,
and it doesn't avoid the startup overhead of Racket, but hopefully the constants
will get better once we switch over completely to Haskell. After that we can
start optimising the big-O, pre-computing things at compile time, etc.