Only run on 'samples' arrays
This means we can run one bucketing method on some samples, another bucketing
method on the result of that, another bucketing method on the result of that,
and so on (it composes).

When we've bucketed all that we want, the ground truths script will process them
all. Note that such ground-truth-added data cannot be subsequently bucketed.

Maybe we should add some assertions for that in the Python?