Always use python.withPackages, even if we have no deps, to force PYTHONHOME
It looks like there's a difference between "${python}/bin/python" and
"${python.withPackages (p: [])}/bin/python" (and likewise for python3), in that
the latter is a wrapper script which sets 'PYTHONHOME' before running the actual
python executable (or subsequent wrapper, whatever).

The problem is, this 'PYTHONHOME' env var will be inherited by any subprocesses
called from that Python script. If a script running with a "bare" python binary
invokes another Python script as a subprocess, that should be fine. If a script
running with a "wrapped" python binary invokes another Python script as a
subprocess, that other script will inherit the 'PYTHONPATH' env var.

If that subprocess script uses a "wrapped" binary then it should be OK, since
the inherited 'PYTHONHOME' will be overwritten by the wrapped binary's. If not,
it may alter the behaviour of the Python interpreter.

I think this we can get away unscathed if both scripts are using the same python
version (one with, and one without, the wrapper). Yet if they differ, this
inherited env var will prevent the subprocess's Python interpreter from finding
its own modules!

This manifests as error messages like:

    Fatal Python error: Py_Initialize: Unable to get the locale encoding
    ImportError: No module named 'encodings'

    Current thread 0x00007f108baec700 (most recent call first):

This commit replaces all dependencies on "raw" Python interpreters with
"wrapped" ones. This way, they should all get their own 'PYTHONHOME' set, and
hence won't use the inherited versions.