Speed up GetGroundTruths by avoiding decoding
Our input names are encoded (e.g. 'globalXXX') but our "DB" of theorem deps are
not (e.g. 'plus'), To compare the two, we need to get them into the same format.

We were decoding the input names at runtime. This is *very* wasteful, since
there are a huge number of names in our input.

We now encode the DB names instead. This is better since (a) there's fewer of
them and (b) it's done at compile time.

Encoding also seems to be slightly faster than decoding, since it's a
straightforward concatMap (in fact, we use the appendable nature of ShowS to our
advantage!); decoding needs to pop two Chars at a time, which makes for a less
obvious (and presumably optimisable) loop structure.