Profile on fewer, larger samples
If we're running on samples from size 1 to 100, most of them are going to have
tens of definitions. This commit swaps out the 'heftySamples' used for profiling
GetGroundTruths so that it has larger samples (although fewer overall, to save
time).

We also add buckets to these samples using a separate derivation, so the result
can be re-used across builds.

The file profile-results/6_resampled.prof is the new baseline, using the new
test input and without changing anything about the implementation. A third of
the time and memory is spent walking along the input; another third is spent
checking for subsets.