Use BitSet instead of AscendingList
A BitSet uses the bits of a Word (or Integer, for arbitrary lengths) to store
whether or not each value of an Enum type is present or not. Since we only have
a few hundred Names, which we're already assigning to Ints via a HashMap, we
only need a handful of Words to store one bit for each Name.

Checking for subsets is very fast for BitSets: we do a bitwise OR of the two
Words, then an equality check with the second. Semantically this is testing
whether the union the two contains any more values than the second.

An extra advantage of using BitSets compared to AscendingLists is that they
automatically discard duplicates (setting a bit is idempotent) and there is no
need to do any sorting. This makes some of our functions obsolete, so I've
deleted them (and their tests).

I've put the profiling info in profile-results/7_bitset.prof. The main takeaways
are:

  - Took 4 seconds, down from 5
  - Allocated 800M, up from 600M
  - Roughly equal time spent parsing and theorem-finding (~40% each)
  - About 24% of the time is spent looking for matches